Improving Error Handling
Earlier error handling was handled by ApplicationError which was
mapped into a ApiError in the API layer, now they are separated.
ApplicationError maps to ApiError, most of the time to the generic
InternalError but sometimes also specific errors such as NotFound
to give valid user feedback. Added ApiError for BadRequest